Exception Handler delete

+++ src/main/java/com/vitoreduardo/course/services/UserService.java
@@ -2,8 +2,11 @@ package com.vitoreduardo.course.services;
 
 import com.vitoreduardo.course.entities.User;
 import com.vitoreduardo.course.repositories.UserRepository;
+import com.vitoreduardo.course.services.exceptions.DatabaseException;
 import com.vitoreduardo.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,13 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+
+            repository.deleteById(id);
+        } catch(DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
